Reference

uESTplanar.KernelConvolution

uESTplanar.KernelConvolution

Base.size

uESTplanar.K

uESTplanar.round_pow2

uESTplanar.∂xK

uESTplanar.∂yK

uESTplanar.∂zK

convolutions

uESTplanar.KernelConvolution — Type

KernelConvolution{T, M}

Convolution kernel type contains a grid information and the kernel convolution matrix. Instances of KernelConvolution can be used to convolve 2-dimensional functions.

The convolution is zero-padded and the type also allocates a buffer allowing the kernel transform to be applied many times with minimal memory usage.

Types

T<:AbstactFloat floating point precision type

M::Int dimension of kernel function. For scalar kernals M =1, for vector kernels M = 3

Properties

xs::StepRangeLen{T} grid of x values in cartesian grid

ys::StepRangeLen{T} grid of y values in cartesian grid

kermat::Array{Complex{T}} M-dimensional Fourier space kernel matrix

buffer::Array{Complex{T}} M-dimensional buffer

uESTplanar.KernelConvolution — Method

KernelConvolution{T}(xs, ys, kernel, M=1)

Construct a kernel convolution over grid xs, ys using kernel(x, y) function. The M specifies the dimension of the kernel function output.

For example, if the kernel is scalar function then kernel(x, y) should return a scalar value and M = 1. If the kernel is a vector function where kernel(x,y) outputs a vector of length n then M = n.

Examples

Defining a 2-dimensional kernel function

julia> kernel = (x, y) -> [cos(x), sin(y)];
julia> xs = range(0, 1, 128)
julia> KernelConvolution{Float64}(xs, xs, kernel, 2);

Base.size — Method

size(kconv::KernelConvolution{T, M})

Return the size of the convolution grid

uESTplanar.round_pow2 — Function

round_pow2(n::Integer)

Round an integer to the nearest power of two.

Examples

julia> round_pow2(4)
4
julia> round_pow2(5)
4

round_pow2(xs::AbstractRange)

Modify the step size so the range length is the nearest power of two.

Examples

julia> xs = 1:7;
julia> round_pow2(xs)
1:1:8

kernels

uESTplanar.K — Function

K(x, y, z)

Planar electrostatic trap kernel in cartesian coordinates where 0 < z < 1 and x, y are real valued horizontal coordinates. The bottom and top plates of the trap are located at z = 0 and z = 1, respectively.

Useful for computing the electrostatic field via convolution.

See also ∂xK, ∂yK, ∂zK.

K(ρ, z)

Planar electrostatic trap kernel in cylindrical coordinates where 0 < ρ, 0 < z < 1.

uESTplanar.∂xK — Function

∂xK(x, y, z)

Partial derivative wrt. x of planar electrostatic trap kernel in cartersian coordinates where 0 < z < 1.

See also K

uESTplanar.∂yK — Function

∂yK(x, y, z)

Partial derivative wrt. y of planar electrostatic trap kernel in cartersian coordinates where 0 < z < 1.

See also K

uESTplanar.∂zK — Function

∂zK(x, y, z)

Partial derivative wrt. z of planar electrostatic trap kernel in cartersian coordinates where 0 < z < 1.

See also K