Reference

uESTplanar.KernelConvolution

uESTplanar.K

uESTplanar.∂xK

uESTplanar.∂yK

uESTplanar.∂zK

convolutions

uESTplanar.KernelConvolution — Type

KernelConvolution{T}

Convolution kernel computed over a fixed 2-dimensional grid xs, ys. Can be used to convolve 2-dimensional functions.

The convolution is zero-padded.

Properties

xs::StepRangeLen grid of x values in cartesian grid

ys::StepRangeLen grid of y values in cartesian grid

kermatt

kernels

uESTplanar.K — Function

K(x, y, z)

Planar electrostatic trap kernel in cartesian coordinates where 0 < z < 1 and x, y are real valued horizontal coordinates. The bottom and top plates of the trap are located at z = 0 and z = 1, respectively.

Useful for computing the electrostatic field via convolution.

See also ∂xK, ∂yK, ∂zK.

K(ρ, z)

Planar electrostatic trap kernel in cylindrical coordinates where 0 < ρ, 0 < z < 1.

uESTplanar.∂xK — Function

∂xK(x, y, z)

Partial derivative wrt. x of planar electrostatic trap kernel in cartersian coordinates where 0 < z < 1.

See also K

uESTplanar.∂yK — Function

∂yK(x, y, z)

Partial derivative wrt. y of planar electrostatic trap kernel in cartersian coordinates where 0 < z < 1.

See also K

uESTplanar.∂zK — Function

∂zK(x, y, z)

Partial derivative wrt. z of planar electrostatic trap kernel in cartersian coordinates where 0 < z < 1.

See also K